Move the greeting command and data to separate files
